requirements.txt
django==5.0.9
psycopg2-binary==2.9.9
gunicorn==22.0.0
requests==2.32.3
pydantic==2.8.2
faker==26.0.0
django-admin-interface==0.28.8
djangorestframework==3.15.2
psutil==6.0.0
sentry-sdk[django]==2.10.0
boto3==1.34.145
pytest==8.3.1
pytest-django==4.8.0
openai==1.37.0